requirements.txt
bentoml==1.4.3
pandas==2.0.3
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.4
pytest-cov==5.0.0
requests==2.32.3
scikit-learn==1.3.2
matplotlib==3.7.5 
pyjwt==2.9.0
